Change datatype from decimal to integer for simplicity

+++ db/migrate/20190715155721_create_warehouses_products.rb
@@ -3,8 +3,8 @@ class CreateWarehousesProducts < ActiveRecord::Migration[5.2]
     create_table :warehouses_products do |t|
       t.integer :product_id, index: true
       t.integer :warehouse_id, index: true
-      t.decimal :item_count
-      t.decimal :low_item_threshold
+      t.integer :item_count
+      t.integer :low_item_threshold
     end
   end
 end
